detectors: generate man pages for modules
IMPORTANT: include of aminclude.am must have been moved to the bottom of
Makefile.am!  If the include is on the top as it used to be, new target
for man does not know about PROGRAMS and SOURCES.

NOTE: we assume there is just one generated program at once.  Having
more bin_PROGRAMS would probably not generate correct man page.