smtp_spam_detector: BUGFIX exec python script instead of fork
The bash script smtp_spam_detector executed the smtp_daemon script
using a standard way - fork() & exec().
This behavior causes that supervisor cannot find a correct service socket (IFC).

The patch fixes this issue because the executed python script runs in the
same PID as the original shell script, so the service IFC is the same.