bl_downloader: fixed writing of output files, described format in README
Generation of output files rewritten and fixed to not contain "ports" part if there is not port information.

NOTICE: The ports extension of the file format wasn't documented anywhere, so I reconstructed its probable definition from the parser code in ipblacklistfilter.cpp. It seems the files generated by bl_downloader can now be successfully read by ipblacklistfilter, but it's not thoroughly tested.